# api/openapi/stackkits-v1.yaml
openapi: "3.1.0"
info:
  title: kombify StackKits API
  version: "1.0.0"
  description: |
    kombify StackKits HTTP API — pre-packaged homelab infrastructure templates
    with CUE schema validation and OpenTofu generation. Designed for consumption
    by the kombify Stack control plane and AI agents via the kombify Gateway.
  contact:
    name: kombify
    url: https://kombify.io
  license:
    name: MIT
servers:
  - url: http://localhost:8082
    description: Local development
  - url: https://api.kombify.io/stackkits
    description: Production (via Cloudflare Edge)
paths:
...
  /api/v1/stackkits/{name}:
    get:
      operationId: getStackKit
      summary: Get full StackKit details
      tags: [Catalog]
      parameters:
        - $ref: "#/components/parameters/StackKitName"
      responses:
        "200":
          description: Full StackKit configuration
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        $ref: "#/components/schemas/StackKit"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/validate:
    post:
      operationId: validateSpec
      summary: Validate a stack-spec against its StackKit schema
      description: >-
        Canonical StackSpec v2 is validated against the embedded CUE authority and produces spec-only evidence. On native v0.7, StackSpec v1 remains accepted only for read-only validation and migration diagnostics; it is never operationally admitted.
      tags: [Validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/ArchitectureV2StackSpec"
                - $ref: "#/components/schemas/ArchitectureV1DiagnosticsInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        oneOf:
                          - $ref: "#/components/schemas/ArchitectureV2SpecValidation"
                          - $ref: "#/components/schemas/ArchitectureV1ReadOnlyValidation"
                          - $ref: "#/components/schemas/LegacyValidationResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /api/v1/generate/preview:
    post:
      operationId: generatePreview
      summary: Preview generated infrastructure without writing files
      description: >-
        Exact-v0.6 compatibility endpoint only. Native v0.7 does not advertise this capability and never treats spec-only input as generation-ready.
      deprecated: true
      tags: [Generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Preview of generated output
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - properties:
                      data:
                        $ref: "#/components/schemas/PreviewResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
components:
  securitySchemes:
    StackKitsAPIKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Required by production deployments; may be disabled only for explicit local development.
...
  parameters:
    StackKitName:
      name: name
      in: path
      required: true
      description: StackKit name (e.g., basement-kit)
      schema:
        type: string
        pattern: "^[a-z0-9][a-z0-9-]*[a-z0-9]$"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [data]
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ResponseMeta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: integer
            message:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ArchitectureV2StackSpec:
      type: object
      description: |
        Canonical StackSpec v2 intent. Unknown top-level fields are rejected;
        the embedded CUE authority remains authoritative for all deeper
        cross-field constraints and defaults.
      additionalProperties: false
      required:
        - apiVersion
        - kind
        - metadata
        - kit
        - install
        - generation
        - network
        - sites
        - nodes
        - controlPlane
        - capabilities
        - availability
        - partitionPolicy
      properties:
        apiVersion:
          type: string
          const: stackkit/v2alpha1
        kind:
          type: string
          const: StackSpec
        metadata:
          type: object
          additionalProperties: false
          required: [name]
          properties:
            name:
              $ref: "#/components/schemas/ArchitectureV2ContractID"
            stackId:
              $ref: "#/components/schemas/ArchitectureV2ContractID"
            fleetRef:
              $ref: "#/components/schemas/ArchitectureV2ContractID"
        source:
          type: object
          additionalProperties: false
          required: [kind]
          properties:
            kind:
              type: string
              enum: [native-v2, migrated-v1, registry]
            ref:
              type: string
              pattern: "^[^\\s]+$"
            migration:
              type: object
              additionalProperties: true
        kit:
          type: object
          additionalProperties: false
          required: [slug]
          properties:
            slug:
              $ref: "#/components/schemas/ArchitectureV2KitProfile"
            version:
              type: string
        install:
          type: object
        generation:
          type: object
        system:
          type: object
        storage:
          type: object
        network:
          type: object
        container:
          type: object
        sites:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ArchitectureV2SiteSpec"
        nodes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ArchitectureV2NodeSpec"
        controlPlane:
          type: object
        capabilities:
          type: object
          additionalProperties: false
          required: [enable, disable]
          properties:
            enable:
              type: array
              items:
                $ref: "#/components/schemas/ArchitectureV2ContractID"
            disable:
              type: array
              items:
                $ref: "#/components/schemas/ArchitectureV2ContractID"
            config:
              type: object
              additionalProperties:
                type: object
                additionalProperties: false
                properties:
                  providerRef:
                    $ref: "#/components/schemas/ArchitectureV2ContractID"
                  settings:
                    type: object
                    additionalProperties: true
                  secretRefs:
                    $ref: "#/components/schemas/ArchitectureV2SecretRefs"
        addons:
          type: object
          additionalProperties:
            type: object
        access:
          type: object
          additionalProperties:
            type: object
        lanDiscovery:
          $ref: "#/components/schemas/ArchitectureV2LANDiscoveryIntent"
        bridge:
          type: object
        availability:
          type: object
        deviceEnrollment:
          type: object
        partitionPolicy:
          type: object
        workloads:
          type: object
          additionalProperties:
            type: object
        modules:
          type: object
          additionalProperties:
            type: object
        routes:
          type: object
          additionalProperties:
            type: object
        data:
          type: object
    ArchitectureV2LANDiscoveryIntent:
      type: object
      description: >-
        Explicit allowlist of local Home routes that may be advertised on the LAN. An absent or empty list never infers advertisements from routes.
      additionalProperties: false
      properties:
        advertiseRouteRefs:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/ArchitectureV2ContractID"
...
    ArchitectureV2SiteSpec:
      type: object
      description: >-
        Provider-neutral placement intent. A site identifies trust and failure domains; it never identifies a server provider, account, region, datacenter, native resource, credential, or management address.
      additionalProperties: false
      required: [id, kind, failureDomain]
      properties:
        id:
          $ref: "#/components/schemas/ArchitectureV2ContractID"
        kind:
          type: string
          enum: [home, cloud]
        name:
          type: string
        failureDomain:
          type: string
          minLength: 1
        network:
          type: object
          additionalProperties: false
          properties:
            serviceCIDRs:
              type: array
              items:
                type: string
            storageCIDRs:
              type: array
              items:
                type: string
        labels:
          type: object
          additionalProperties:
            type: string
...
    ArchitectureV1DiagnosticsInput:
      type: object
      description: |
        Lossless legacy StackSpec v1 input accepted only to produce migration
        diagnostics. Fields are deliberately not projected or discarded at the
        HTTP boundary, and this branch can never match a StackSpec v2 header.
      additionalProperties: true
      oneOf:
        - required: [apiVersion]
          properties:
            apiVersion:
              type: string
              const: stackkit/v1
        - not:
            required: [apiVersion]
...
    ArchitectureV2KitProfile:
      type: string
      enum: [basement-kit, cloud-kit]
...
    ArchitectureV2ContractID:
      type: string
      pattern: "^[a-z][a-z0-9-]*$"
...
    ArchitectureV2ContentHash:
      type: string
      pattern: "^sha256:[a-f0-9]{64}$"
...
    ArchitectureV2SecretRefs:
      type: object
      additionalProperties:
        type: string
        pattern: "^(secret|vault|doppler|techstack)://[^\\s]+$"
...
    StackKit:
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/StackKitMetadata"
        supportedOS:
          type: array
          items:
            type: string
        requirements:
          $ref: "#/components/schemas/Requirements"
        modes:
          $ref: "#/components/schemas/Modes"
        features:
          $ref: "#/components/schemas/Features"
    StackKitMetadata:
      type: object
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        name:
          type: string
        version:
          type: string
        displayName:
          type: string
        description:
          type: string
        author:
          type: string
        license:
          type: string
        homepage:
          type: string
        repository:
          type: string
        tags:
          type: array
          items:
            type: string
    Requirements:
      type: object
      properties:
        minimum:
          $ref: "#/components/schemas/ResourceSpec"
        recommended:
          $ref: "#/components/schemas/ResourceSpec"
    ResourceSpec:
      type: object
      properties:
        cpu:
          type: integer
        ram:
          type: integer
          description: RAM in GB
        disk:
          type: integer
          description: Disk in GB
    Modes:
      type: object
      properties:
        simple:
          $ref: "#/components/schemas/ModeSpec"
        advanced:
          $ref: "#/components/schemas/ModeSpec"
...
    Features:
      type: object
      properties:
        multiNode:
          type: boolean
        vpnOverlay:
          type: boolean
        publicAccess:
          type: boolean
...
    ValidationResult:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
    LegacyValidationResult:
      allOf:
        - $ref: "#/components/schemas/ValidationResult"
        - type: object
          description: Exact-v0.6 validation response without a version discriminator.
          not:
            required: [source_version]
    ArchitectureV2SpecValidation:
      type: object
      additionalProperties: false
      required: [valid, source_version, stackkit, spec_hash]
      properties:
        valid:
          type: boolean
          const: true
        source_version:
          type: string
          const: v2alpha1
        stackkit:
          $ref: "#/components/schemas/ArchitectureV2KitProfile"
        spec_hash:
          $ref: "#/components/schemas/ArchitectureV2ContentHash"
    ArchitectureV1ReadOnlyValidation:
      allOf:
        - $ref: "#/components/schemas/ValidationResult"
        - type: object
          required: [source_version, stackkit, operational, migration_required, validation_scope, migration_report]
          properties:
            source_version:
              type: string
              const: v1
            stackkit:
              type: string
            operational:
              type: boolean
              const: false
            migration_required:
              type: boolean
              const: true
            validation_scope:
              type: string
              const: legacy-read-only
            migration_report:
              $ref: "#/components/schemas/ArchitectureV2MigrationReport"
...
    GenerateRequest:
      type: object
      required: [spec]
      properties:
        spec:
          $ref: "#/components/schemas/StackSpec"
        outputDir:
          type: string
          description: Custom output directory (server-side only)
...
    PreviewResult:
      type: object
      properties:
        tfvars:
          type: object
        preview:
          type: boolean
          const: true
        stackkit:
          $ref: "#/components/schemas/StackKitSummary"
...
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Validation or generation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...